Update function _getitem_
The function is fixed such that it will fetch the 1D numpy array of tag's vector, if a single string or input is passed. If a list is given as an argument, then 2D numpy array of tags' vector is returned. And if the passed input is found out to be none of the above mentioned types, it will throw an error.